+++ src/main/scala/com/twitter/scalding/commons/source/PailSource.scala
@@ -21,9 +21,9 @@ import backtype.hadoop.pail.{Pail, PailStructure}
 import cascading.pipe.Pipe
 import cascading.scheme.Scheme
 import cascading.tap.Tap
+import com.twitter.bijection.Bijection
 import com.twitter.chill.MeatLocker
 import com.twitter.scalding._
-import com.twitter.util.Codec
 import java.util.{ List => JList }
 import org.apache.hadoop.mapred.{ JobConf, OutputCollector, RecordReader }
 import scala.collection.JavaConverters._
@@ -35,7 +35,7 @@ import scala.collection.JavaConverters._
  * each tuple.
  */
 
-// This pail structure pulls in an implicit codec to handle
+// This pail structure pulls in an implicit bijection to handle
 // serialization. targetFn takes an instance of T and returns a list
 // of "path components". Pail joins these components with
 // File.separator and sinks the instance of T into the pail at that
@@ -46,14 +46,14 @@ import scala.collection.JavaConverters._
 // targetFn. I don't know that there's a good way to do this check at
 // compile time.
 
-class CodecPailStructure[T](targetFn: (T) => List[String], validator: (List[String]) => Boolean)
-(@transient implicit val codec: Codec[T, Array[Byte]], manifest: Manifest[T])
+class CodecPailStructure[T](targetFn: T => List[String], validator: List[String] => Boolean)
+(implicit @transient bijection: Bijection[T, Array[Byte]], manifest: Manifest[T])
 extends PailStructure[T] {
-  val safeCodec = new MeatLocker(codec)
+  val codecBox = MeatLocker(bijection)
   override def isValidTarget(paths: String*): Boolean = validator(paths.toList)
   override def getTarget(obj: T): JList[String] = targetFn(obj).toList.asJava
-  override def serialize(obj: T): Array[Byte] = safeCodec.get.encode(obj)
-  override def deserialize(bytes: Array[Byte]): T = safeCodec.get.decode(bytes)
+  override def serialize(obj: T): Array[Byte] = codecBox.get.apply(obj)
+  override def deserialize(bytes: Array[Byte]): T = codecBox.get.invert(bytes)
   override val getType = manifest.erasure.asInstanceOf[Class[T]]
 }
 
@@ -68,11 +68,11 @@ object PailSource {
   // A PailSource can also build its structure on the fly from a
   // couple of functions.
   def apply[T](rootPath: String, targetFn: (T) => List[String], validator: (List[String]) => Boolean)
-  (implicit codec: Codec[T,Array[Byte]], manifest: Manifest[T]) =
+  (implicit bijection: Bijection[T, Array[Byte]], manifest: Manifest[T]) =
     new PailSource(rootPath, new CodecPailStructure[T](targetFn, validator), null)
 
   def apply[T](rootPath: String, targetFn: (T) => List[String], validator: (List[String]) => Boolean, subPaths: Array[List[String]])
-  (implicit codec: Codec[T,Array[Byte]], manifest: Manifest[T]) =
+  (implicit bijection: Bijection[T,Array[Byte]], manifest: Manifest[T]) =
     new PailSource(rootPath, new CodecPailStructure[T](targetFn, validator), subPaths)
 }